+++ docs/source/conf.py
@@ -16,7 +16,7 @@ sys.path.insert(0, os.path.abspath("../.."))
 project = 'fastlowess'
 copyright = '2025, Amir Valizadeh'
 author = 'Amir Valizadeh'
-release = '0.2.0'
+release = '0.4.0'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration